Move `validate_link_label` to `aiida.common.links`
The function is general enough to justify it being defined in `aiida.common`
alongside with the `LinkType` enum. In addition, this prevents cyclic
imports problems between the `aiida.engine` and `aiida.orm` modules.
