+++ aiida/backends/tests/__init__.py
@@ -73,6 +73,7 @@ DB_TEST_LIST = {
         'common.extendeddicts': ['aiida.backends.tests.common.test_extendeddicts'],
         'common.folders': ['aiida.backends.tests.common.test_folders'],
         'common.hashing': ['aiida.backends.tests.common.test_hashing'],
+        'common.links': ['aiida.backends.tests.common.test_links'],
         'common.logging': ['aiida.backends.tests.common.test_logging'],
         'common.serialize': ['aiida.backends.tests.common.test_serialize'],
         'common.timezone': ['aiida.backends.tests.common.test_timezone'],
@@ -124,7 +125,6 @@ DB_TEST_LIST = {
         'orm.node': ['aiida.backends.tests.orm.node.test_node'],
         'orm.utils.calcjob': ['aiida.backends.tests.orm.utils.test_calcjob'],
         'orm.utils.node': ['aiida.backends.tests.orm.utils.test_node'],
-        'orm.utils.links': ['aiida.backends.tests.orm.utils.test_links'],
         'orm.utils.loaders': ['aiida.backends.tests.orm.utils.test_loaders'],
         'orm.utils.repository': ['aiida.backends.tests.orm.utils.test_repository'],
         'parsers.parser': ['aiida.backends.tests.parsers.test_parser'],
